CDAP-9474 separate spark core modules
Creating separate spark-core modules for spark1 and spark2.
Most of the existing code from spark-core was moved to
spark-core-base. From there, cdap-api-spark was removed as a
dependency, and any class that uses an api class was moved into
both cdap-spark-core and cdap-spark-core2_2.11.

Each Spark core module has its own SparkProgramRuntimeProvider
that in turn instantiates its own SparkProgramRunner that in
turn sets its own SparkMainWrapper class. Each
ProgramRuntimeProvider is also annotated with a new SparkCompat
property so that only one of them will be used at runtime.

SparkClassLoader was refactored into a base class with a Spark1
and Spark2 implementations so that its static method to get
the classloader could be divided between the Spark1 and Spark2
main wrappers.

Classloaders from Spark1.2 were moved to only the cdap-spark-core
module.